% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessingGEO.R
\name{accessingGEO}
\alias{accessingGEO}
\title{This will download a 20 Mb}
\usage{
accessingGEO(accessNumber)
}
\arguments{
\item{accessNumber}{a character string to download}
}
\value{
expression set
}
\description{
This will download a 20 Mb
}

